changed hpp files into h files
