+++ applications/GeoMechanicsApplication/custom_conditions/Pw_condition.cpp
@@ -12,7 +12,7 @@
 //
 
 // Application includes
-#include "custom_conditions/Pw_condition.hpp"
+#include "custom_conditions/Pw_condition.h"
 #include "custom_utilities/dof_utilities.h"
 
 namespace Kratos
@@ -122,6 +122,18 @@ template <unsigned int TDim, unsigned int TNumNodes>
 Condition::DofsVectorType PwCondition<TDim, TNumNodes>::GetDofs() const
 {
     return Geo::DofUtilities::ExtractDofsFromNodes(GetGeometry(), WATER_PRESSURE);
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void PwCondition<TDim, TNumNodes>::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void PwCondition<TDim, TNumNodes>::load(Serializer& rSerializer)
+{
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
 }
 
 template <unsigned int TDim, unsigned int TNumNodes>
