+++ applications/GeoMechanicsApplication/custom_conditions/Pw_normal_flux_condition.h
@@ -19,7 +19,7 @@
 #include "includes/serializer.h"
 
 // Application includes
-#include "custom_conditions/Pw_condition.hpp"
+#include "custom_conditions/Pw_condition.h"
 #include "custom_utilities/element_utilities.hpp"
 
 namespace Kratos
@@ -36,16 +36,13 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    // Default constructor
     PwNormalFluxCondition() : PwCondition<TDim, TNumNodes>() {}
 
-    // Constructor 1
     PwNormalFluxCondition(IndexType NewId, GeometryType::Pointer pGeometry)
         : PwCondition<TDim, TNumNodes>(NewId, pGeometry)
     {
     }
 
-    // Constructor 2
     PwNormalFluxCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
         : PwCondition<TDim, TNumNodes>(NewId, pGeometry, pProperties)
     {
@@ -62,16 +59,9 @@ protected:
 
 private:
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
-    }
 }; // class PwNormalFluxCondition.
 
 } // namespace Kratos.
