+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_face_load_condition.cpp
@@ -13,7 +13,7 @@
 //
 
 // Application includes
-#include "custom_conditions/U_Pw_face_load_condition.hpp"
+#include "custom_conditions/U_Pw_face_load_condition.h"
 #include "custom_utilities/condition_utilities.hpp"
 #include "custom_utilities/element_utilities.hpp"
 
@@ -72,6 +72,18 @@ template <unsigned int TDim, unsigned int TNumNodes>
 std::string UPwFaceLoadCondition<TDim, TNumNodes>::Info() const
 {
     return "UPwFaceLoadCondition";
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwFaceLoadCondition<TDim, TNumNodes>::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwFaceLoadCondition<TDim, TNumNodes>::load(Serializer& rSerializer)
+{
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
 }
 
 template class UPwFaceLoadCondition<2, 2>;
