+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_face_load_interface_condition.cpp
@@ -13,7 +13,7 @@
 //
 
 // Application includes
-#include "custom_conditions/U_Pw_face_load_interface_condition.hpp"
+#include "custom_conditions/U_Pw_face_load_interface_condition.h"
 #include "custom_utilities/element_utilities.hpp"
 
 namespace Kratos
@@ -300,6 +300,18 @@ template <unsigned int TDim, unsigned int TNumNodes>
 std::string UPwFaceLoadInterfaceCondition<TDim, TNumNodes>::Info() const
 {
     return "UPwFaceLoadInterfaceCondition";
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwFaceLoadInterfaceCondition<TDim, TNumNodes>::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwFaceLoadInterfaceCondition<TDim, TNumNodes>::load(Serializer& rSerializer)
+{
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
 }
 
 template class UPwFaceLoadInterfaceCondition<2, 2>;
