+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_face_load_interface_condition.h
@@ -18,7 +18,7 @@
 #include "includes/serializer.h"
 
 // Application includes
-#include "custom_conditions/U_Pw_condition.hpp"
+#include "custom_conditions/U_Pw_condition.h"
 #include "custom_utilities/condition_utilities.hpp"
 #include "custom_utilities/interface_element_utilities.h"
 #include "geo_mechanics_application_variables.h"
@@ -85,19 +85,10 @@ protected:
 private:
     Vector mInitialGap;
 
-    // Serialization
-
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
-    }
 }; // class UPwFaceLoadInterfaceCondition.
 
 } // namespace Kratos.
