+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_normal_face_load_condition.cpp
@@ -13,7 +13,7 @@
 //
 
 // Application includes
-#include "custom_conditions/U_Pw_normal_face_load_condition.hpp"
+#include "custom_conditions/U_Pw_normal_face_load_condition.h"
 #include "custom_utilities/element_utilities.hpp"
 #include "custom_utilities/variables_utilities.hpp"
 
@@ -125,6 +125,17 @@ std::string UPwNormalFaceLoadCondition<TDim, TNumNodes>::Info() const
     return "UPwNormalFaceLoadCondition";
 }
 
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwNormalFaceLoadCondition<TDim, TNumNodes>::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwNormalFaceLoadCondition<TDim, TNumNodes>::load(Serializer& rSerializer)
+{
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
+}
 template class UPwNormalFaceLoadCondition<2, 2>;
 template class UPwNormalFaceLoadCondition<2, 3>;
 template class UPwNormalFaceLoadCondition<2, 4>;
