+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_normal_face_load_condition.h
@@ -18,7 +18,7 @@
 #include "includes/serializer.h"
 
 // Application includes
-#include "custom_conditions/U_Pw_condition.hpp"
+#include "custom_conditions/U_Pw_condition.h"
 #include "custom_utilities/condition_utilities.hpp"
 #include "geo_mechanics_application_variables.h"
 
@@ -36,16 +36,13 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    // Default constructor
     UPwNormalFaceLoadCondition() : UPwCondition<TDim, TNumNodes>() {}
 
-    // Constructor 1
     UPwNormalFaceLoadCondition(IndexType NewId, GeometryType::Pointer pGeometry)
         : UPwCondition<TDim, TNumNodes>(NewId, pGeometry)
     {
     }
 
-    // Constructor 2
     UPwNormalFaceLoadCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
         : UPwCondition<TDim, TNumNodes>(NewId, pGeometry, pProperties)
     {
@@ -63,8 +60,6 @@ protected:
         array_1d<double, TNumNodes> TangentialStressVector;
     };
 
-    // Member Variables
-
     void CalculateRHS(Vector& rRightHandSideVector, const ProcessInfo& CurrentProcessInfo) override;
 
     void InitializeConditionVariables(NormalFaceLoadVariables& rVariables, const GeometryType& Geom);
@@ -79,19 +74,10 @@ protected:
                                                    const GeometryType::IntegrationPointsArrayType& IntegrationPoints) const;
 
 private:
-    // Serialization
-
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
-    }
 }; // class UPwNormalFaceLoadCondition.
 
 } // namespace Kratos.
