+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_normal_flux_interface_condition.cpp
@@ -13,7 +13,7 @@
 //
 
 // Application includes
-#include "custom_conditions/U_Pw_normal_flux_interface_condition.hpp"
+#include "custom_conditions/U_Pw_normal_flux_interface_condition.h"
 
 namespace Kratos
 {
@@ -97,6 +97,17 @@ std::string UPwNormalFluxInterfaceCondition<TDim, TNumNodes>::Info() const
     return "UPwNormalFluxInterfaceCondition";
 }
 
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwNormalFluxInterfaceCondition<TDim, TNumNodes>::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
+}
+
+template <unsigned int TDim, unsigned int TNumNodes>
+void UPwNormalFluxInterfaceCondition<TDim, TNumNodes>::load(Serializer& rSerializer)
+{
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
+}
 template class UPwNormalFluxInterfaceCondition<2, 2>;
 template class UPwNormalFluxInterfaceCondition<3, 4>;
