+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_normal_flux_interface_condition.h
@@ -18,36 +18,34 @@
 #include "includes/serializer.h"
 
 // Application includes
-#include "custom_conditions/U_Pw_condition.hpp"
-#include "custom_utilities/condition_utilities.hpp"
-#include "geo_mechanics_application_variables.h"
+#include "custom_conditions/U_Pw_condition.h"
+#include "custom_conditions/U_Pw_face_load_interface_condition.h"
+#include "custom_utilities/element_utilities.hpp"
 
 namespace Kratos
 {
 
 template <unsigned int TDim, unsigned int TNumNodes>
-class KRATOS_API(GEO_MECHANICS_APPLICATION) UPwFaceLoadCondition : public UPwCondition<TDim, TNumNodes>
+class KRATOS_API(GEO_MECHANICS_APPLICATION) UPwNormalFluxInterfaceCondition
+    : public UPwFaceLoadInterfaceCondition<TDim, TNumNodes>
 {
 public:
-    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(UPwFaceLoadCondition);
+    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(UPwNormalFluxInterfaceCondition);
 
     using IndexType      = std::size_t;
     using PropertiesType = Properties;
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    // Default constructor
-    UPwFaceLoadCondition() : UPwCondition<TDim, TNumNodes>() {}
+    UPwNormalFluxInterfaceCondition() : UPwFaceLoadInterfaceCondition<TDim, TNumNodes>() {}
 
-    // Constructor 1
-    UPwFaceLoadCondition(IndexType NewId, GeometryType::Pointer pGeometry)
-        : UPwCondition<TDim, TNumNodes>(NewId, pGeometry)
+    UPwNormalFluxInterfaceCondition(IndexType NewId, GeometryType::Pointer pGeometry)
+        : UPwFaceLoadInterfaceCondition<TDim, TNumNodes>(NewId, pGeometry)
     {
     }
 
-    // Constructor 2
-    UPwFaceLoadCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
-        : UPwCondition<TDim, TNumNodes>(NewId, pGeometry, pProperties)
+    UPwNormalFluxInterfaceCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
+        : UPwFaceLoadInterfaceCondition<TDim, TNumNodes>(NewId, pGeometry, pProperties)
     {
     }
 
@@ -62,16 +60,9 @@ protected:
 
 private:
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
-    }
-}; // class UPwFaceLoadCondition.
+}; // class UPwNormalFluxInterfaceCondition.
 
 } // namespace Kratos.
