--- applications/GeoMechanicsApplication/custom_conditions/U_Pw_normal_flux_interface_condition.hpp
@@ -1,82 +1,0 @@
-// KRATOS___
-//     //   ) )
-//    //         ___      ___
-//   //  ____  //___) ) //   ) )
-//  //    / / //       //   / /
-// ((____/ / ((____   ((___/ /  MECHANICS
-//
-//  License:         geo_mechanics_application/license.txt
-//
-//
-//  Main authors:    Ignasi de Pouplana,
-//                   Vahid Galavi
-//
-
-#pragma once
-
-// Project includes
-#include "includes/serializer.h"
-
-// Application includes
-#include "custom_conditions/U_Pw_condition.hpp"
-#include "custom_conditions/U_Pw_face_load_interface_condition.hpp"
-#include "custom_utilities/element_utilities.hpp"
-
-namespace Kratos
-{
-
-template <unsigned int TDim, unsigned int TNumNodes>
-class KRATOS_API(GEO_MECHANICS_APPLICATION) UPwNormalFluxInterfaceCondition
-    : public UPwFaceLoadInterfaceCondition<TDim, TNumNodes>
-{
-public:
-    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(UPwNormalFluxInterfaceCondition);
-
-    using IndexType      = std::size_t;
-    using PropertiesType = Properties;
-    using GeometryType   = Geometry<Node>;
-    using NodesArrayType = GeometryType::PointsArrayType;
-
-    // Default constructor
-    UPwNormalFluxInterfaceCondition() : UPwFaceLoadInterfaceCondition<TDim, TNumNodes>() {}
-
-    // Constructor 1
-    UPwNormalFluxInterfaceCondition(IndexType NewId, GeometryType::Pointer pGeometry)
-        : UPwFaceLoadInterfaceCondition<TDim, TNumNodes>(NewId, pGeometry)
-    {
-    }
-
-    // Constructor 2
-    UPwNormalFluxInterfaceCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
-        : UPwFaceLoadInterfaceCondition<TDim, TNumNodes>(NewId, pGeometry, pProperties)
-    {
-    }
-
-    Condition::Pointer Create(IndexType               NewId,
-                              NodesArrayType const&   ThisNodes,
-                              PropertiesType::Pointer pProperties) const override;
-
-    std::string Info() const override;
-
-protected:
-    // Member Variables
-
-    void CalculateRHS(Vector& rRightHandSideVector, const ProcessInfo& CurrentProcessInfo) override;
-
-private:
-    // Serialization
-
-    friend class Serializer;
-
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
-    }
-}; // class UPwNormalFluxInterfaceCondition.
-
-} // namespace Kratos.
