+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_normal_lysmer_absorbing_condition.h
@@ -16,8 +16,8 @@
 #include "includes/serializer.h"
 
 // Application includes
-#include "custom_conditions/U_Pw_condition.hpp"
-#include "custom_conditions/U_Pw_face_load_condition.hpp"
+#include "custom_conditions/U_Pw_condition.h"
+#include "custom_conditions/U_Pw_face_load_condition.h"
 #include "custom_utilities/element_utilities.hpp"
 #include "geo_mechanics_application_variables.h"
 
@@ -36,16 +36,13 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    // Default constructor
     UPwLysmerAbsorbingCondition() : UPwFaceLoadCondition<TDim, TNumNodes>() {}
 
-    // Constructor 1
     UPwLysmerAbsorbingCondition(IndexType NewId, GeometryType::Pointer pGeometry)
         : UPwFaceLoadCondition<TDim, TNumNodes>(NewId, pGeometry)
     {
     }
 
-    // Constructor 2
     UPwLysmerAbsorbingCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties)
         : UPwFaceLoadCondition<TDim, TNumNodes>(NewId, pGeometry, pProperties)
     {
@@ -125,8 +122,6 @@ protected:
         DimensionMatrixType CAbsMatrix; // damping part of absorbing matrix;
         DimensionMatrixType KAbsMatrix; // stiffness part of absorbing matrix;
     };
-
-    // Member Variables
 
     /**
      * @brief Adds the condition matrix to the global left hand side
@@ -221,16 +216,9 @@ private:
                                            const ProcessInfo& rCurrentProcessInfo);
 
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, Condition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, Condition)
-    }
 }; // class UPwLysmerAbsorbingCondition.
 
 } // namespace Kratos.
