+++ applications/GeoMechanicsApplication/custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.cpp
@@ -13,7 +13,7 @@
 //
 
 // Project includes
-#include "custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.hpp"
+#include "custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.h"
 #include "custom_utilities/variables_utilities.hpp"
 
 #include <numeric>
@@ -79,6 +79,14 @@ void LineNormalFluidFlux2DDiffOrderCondition::CalculateAndAddConditionForce(Vect
             rVariables.Np[i] * rVariables.ConditionVector[0] * rVariables.IntegrationCoefficient;
     }
 }
+
+void LineNormalFluidFlux2DDiffOrderCondition::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, LineLoad2DDiffOrderCondition)
+}
+
+void LineNormalFluidFlux2DDiffOrderCondition::load(Serializer& rSerializer){
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, LineLoad2DDiffOrderCondition)}
 
 std::string LineNormalFluidFlux2DDiffOrderCondition::Info() const
 {
