+++ applications/GeoMechanicsApplication/custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.h
@@ -15,7 +15,7 @@
 #pragma once
 
 // Project includes
-#include "custom_conditions/general_U_Pw_diff_order_condition.hpp"
+#include "custom_conditions/line_load_2D_diff_order_condition.h"
 #include "includes/serializer.h"
 
 #include "geo_mechanics_application_variables.h"
@@ -23,7 +23,7 @@
 namespace Kratos
 {
 
-class KRATOS_API(GEO_MECHANICS_APPLICATION) LineLoad2DDiffOrderCondition : public GeneralUPwDiffOrderCondition
+class KRATOS_API(GEO_MECHANICS_APPLICATION) LineNormalFluidFlux2DDiffOrderCondition : public LineLoad2DDiffOrderCondition
 {
 public:
     using IndexType      = std::size_t;
@@ -31,16 +31,15 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(LineLoad2DDiffOrderCondition);
+    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(LineNormalFluidFlux2DDiffOrderCondition);
 
-    // Default constructor
-    LineLoad2DDiffOrderCondition();
+    LineNormalFluidFlux2DDiffOrderCondition();
 
-    // Constructor 1
-    LineLoad2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
+    LineNormalFluidFlux2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
 
-    // Constructor 2
-    LineLoad2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties);
+    LineNormalFluidFlux2DDiffOrderCondition(IndexType               NewId,
+                                            GeometryType::Pointer   pGeometry,
+                                            PropertiesType::Pointer pProperties);
 
     Condition::Pointer Create(IndexType               NewId,
                               NodesArrayType const&   ThisNodes,
@@ -50,31 +49,15 @@ public:
     std::string Info() const override;
 
 protected:
-    // Member Variables
-
     void CalculateConditionVector(ConditionVariables& rVariables, unsigned int PointNumber) override;
-
-    double CalculateIntegrationCoefficient(IndexType                          PointNumber,
-                                           const GeometryType::JacobiansType& JContainer,
-                                           const GeometryType::IntegrationPointsArrayType& IntegrationPoints) const override;
 
     void CalculateAndAddConditionForce(Vector& rRightHandSideVector, ConditionVariables& rVariables) override;
 
 private:
-    // Serialization
-
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, GeneralUPwDiffOrderCondition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, GeneralUPwDiffOrderCondition)
-    }
-
-}; // class LineLoad2DDiffOrderCondition.
+}; // class LineNormalFluidFlux2DDiffOrderCondition.
 
 } // namespace Kratos.
