+++ applications/GeoMechanicsApplication/custom_conditions/line_normal_load_2D_diff_order_condition.h
@@ -14,7 +14,8 @@
 
 #pragma once
 
-#include "custom_conditions/surface_load_3D_diff_order_condition.hpp"
+// Project includes
+#include "custom_conditions/line_load_2D_diff_order_condition.h"
 #include "includes/serializer.h"
 
 #include "geo_mechanics_application_variables.h"
@@ -22,7 +23,7 @@
 namespace Kratos
 {
 
-class KRATOS_API(GEO_MECHANICS_APPLICATION) SurfaceNormalFluidFlux3DDiffOrderCondition : public SurfaceLoad3DDiffOrderCondition
+class KRATOS_API(GEO_MECHANICS_APPLICATION) LineNormalLoad2DDiffOrderCondition : public LineLoad2DDiffOrderCondition
 {
 public:
     using IndexType      = std::size_t;
@@ -30,18 +31,13 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(SurfaceNormalFluidFlux3DDiffOrderCondition);
+    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(LineNormalLoad2DDiffOrderCondition);
 
-    // Default constructor
-    SurfaceNormalFluidFlux3DDiffOrderCondition();
+    LineNormalLoad2DDiffOrderCondition();
 
-    // Constructor 1
-    SurfaceNormalFluidFlux3DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
+    LineNormalLoad2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
 
-    // Constructor 2
-    SurfaceNormalFluidFlux3DDiffOrderCondition(IndexType               NewId,
-                                               GeometryType::Pointer   pGeometry,
-                                               PropertiesType::Pointer pProperties);
+    LineNormalLoad2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties);
 
     Condition::Pointer Create(IndexType               NewId,
                               NodesArrayType const&   ThisNodes,
@@ -52,22 +48,17 @@ public:
 
 protected:
     void CalculateConditionVector(ConditionVariables& rVariables, unsigned int PointNumber) override;
+    double CalculateIntegrationCoefficient(IndexType                          PointNumber,
+                                           const GeometryType::JacobiansType& JContainer,
+                                           const GeometryType::IntegrationPointsArrayType& IntegrationPoints) const override;
 
     void CalculateAndAddConditionForce(Vector& rRightHandSideVector, ConditionVariables& rVariables) override;
 
 private:
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, SurfaceLoad3DDiffOrderCondition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, SurfaceLoad3DDiffOrderCondition)
-    }
-
-}; // class SurfaceNormalFluidFlux3DDiffOrderCondition.
+}; // class LineNormalLoad2DDiffOrderCondition.
 
 } // namespace Kratos.
