+++ applications/GeoMechanicsApplication/custom_conditions/surface_load_3D_diff_order_condition.h
@@ -8,13 +8,14 @@
 //  License:         geo_mechanics_application/license.txt
 //
 //
-//  Main authors:    Vahid Galavi
+//  Main authors:    Ignasi de Pouplana,
+//                   Vahid Galavi
 //
 
 #pragma once
 
 // Project includes
-#include "custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.hpp"
+#include "custom_conditions/general_U_Pw_diff_order_condition.h"
 #include "includes/serializer.h"
 
 #include "geo_mechanics_application_variables.h"
@@ -22,8 +23,7 @@
 namespace Kratos
 {
 
-class KRATOS_API(GEO_MECHANICS_APPLICATION) AxisymmetricLineNormalFluidFlux2DDiffOrderCondition
-    : public LineNormalFluidFlux2DDiffOrderCondition
+class KRATOS_API(GEO_MECHANICS_APPLICATION) SurfaceLoad3DDiffOrderCondition : public GeneralUPwDiffOrderCondition
 {
 public:
     using IndexType      = std::size_t;
@@ -31,18 +31,13 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(AxisymmetricLineNormalFluidFlux2DDiffOrderCondition);
+    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(SurfaceLoad3DDiffOrderCondition);
 
-    // Default constructor
-    AxisymmetricLineNormalFluidFlux2DDiffOrderCondition();
+    SurfaceLoad3DDiffOrderCondition();
 
-    // Constructor 1
-    AxisymmetricLineNormalFluidFlux2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
+    SurfaceLoad3DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
 
-    // Constructor 2
-    AxisymmetricLineNormalFluidFlux2DDiffOrderCondition(IndexType               NewId,
-                                                        GeometryType::Pointer   pGeometry,
-                                                        PropertiesType::Pointer pProperties);
+    SurfaceLoad3DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry, PropertiesType::Pointer pProperties);
 
     Condition::Pointer Create(IndexType               NewId,
                               NodesArrayType const&   ThisNodes,
@@ -52,23 +47,19 @@ public:
     std::string Info() const override;
 
 protected:
+    void CalculateConditionVector(ConditionVariables& rVariables, unsigned int PointNumber) override;
+
     double CalculateIntegrationCoefficient(IndexType                          PointNumber,
                                            const GeometryType::JacobiansType& JContainer,
                                            const GeometryType::IntegrationPointsArrayType& IntegrationPoints) const override;
 
+    void CalculateAndAddConditionForce(Vector& rRightHandSideVector, ConditionVariables& rVariables) override;
+
 private:
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, LineNormalFluidFlux2DDiffOrderCondition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, LineNormalFluidFlux2DDiffOrderCondition)
-    }
-
-}; // class AxisymmetricLineNormalFluidFlux2DDiffOrderCondition.
+}; // class SurfaceLoad3DDiffOrderCondition.
 
 } // namespace Kratos.
