+++ applications/GeoMechanicsApplication/custom_conditions/surface_normal_fluid_flux_3D_diff_order_condition.cpp
@@ -13,7 +13,7 @@
 //
 
 // Project includes
-#include "custom_conditions/surface_normal_fluid_flux_3D_diff_order_condition.hpp"
+#include "custom_conditions/surface_normal_fluid_flux_3D_diff_order_condition.h"
 #include "custom_utilities/variables_utilities.hpp"
 
 #include <numeric>
@@ -78,7 +78,14 @@ void SurfaceNormalFluidFlux3DDiffOrderCondition::CalculateAndAddConditionForce(V
     }
 }
 
-std::string SurfaceNormalFluidFlux3DDiffOrderCondition::Info() const
+void SurfaceNormalFluidFlux3DDiffOrderCondition::save(Serializer& rSerializer) const
+{
+    KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, SurfaceLoad3DDiffOrderCondition)
+}
+
+void SurfaceNormalFluidFlux3DDiffOrderCondition::load(Serializer& rSerializer){
+    KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, SurfaceLoad3DDiffOrderCondition)} std::string
+    SurfaceNormalFluidFlux3DDiffOrderCondition::Info() const
 {
     return "SurfaceNormalFluidFlux3DDiffOrderCondition";
 }
