+++ applications/GeoMechanicsApplication/custom_conditions/surface_normal_fluid_flux_3D_diff_order_condition.h
@@ -14,8 +14,7 @@
 
 #pragma once
 
-// Project includes
-#include "custom_conditions/line_load_2D_diff_order_condition.hpp"
+#include "custom_conditions/surface_load_3D_diff_order_condition.h"
 #include "includes/serializer.h"
 
 #include "geo_mechanics_application_variables.h"
@@ -23,7 +22,7 @@
 namespace Kratos
 {
 
-class KRATOS_API(GEO_MECHANICS_APPLICATION) LineNormalFluidFlux2DDiffOrderCondition : public LineLoad2DDiffOrderCondition
+class KRATOS_API(GEO_MECHANICS_APPLICATION) SurfaceNormalFluidFlux3DDiffOrderCondition : public SurfaceLoad3DDiffOrderCondition
 {
 public:
     using IndexType      = std::size_t;
@@ -31,18 +30,15 @@ public:
     using GeometryType   = Geometry<Node>;
     using NodesArrayType = GeometryType::PointsArrayType;
 
-    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(LineNormalFluidFlux2DDiffOrderCondition);
+    KRATOS_CLASS_INTRUSIVE_POINTER_DEFINITION(SurfaceNormalFluidFlux3DDiffOrderCondition);
 
-    // Default constructor
-    LineNormalFluidFlux2DDiffOrderCondition();
+    SurfaceNormalFluidFlux3DDiffOrderCondition();
 
-    // Constructor 1
-    LineNormalFluidFlux2DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
+    SurfaceNormalFluidFlux3DDiffOrderCondition(IndexType NewId, GeometryType::Pointer pGeometry);
 
-    // Constructor 2
-    LineNormalFluidFlux2DDiffOrderCondition(IndexType               NewId,
-                                            GeometryType::Pointer   pGeometry,
-                                            PropertiesType::Pointer pProperties);
+    SurfaceNormalFluidFlux3DDiffOrderCondition(IndexType               NewId,
+                                               GeometryType::Pointer   pGeometry,
+                                               PropertiesType::Pointer pProperties);
 
     Condition::Pointer Create(IndexType               NewId,
                               NodesArrayType const&   ThisNodes,
@@ -52,27 +48,15 @@ public:
     std::string Info() const override;
 
 protected:
-    // Member Variables
-
     void CalculateConditionVector(ConditionVariables& rVariables, unsigned int PointNumber) override;
 
     void CalculateAndAddConditionForce(Vector& rRightHandSideVector, ConditionVariables& rVariables) override;
 
 private:
-    // Serialization
-
     friend class Serializer;
+    void save(Serializer& rSerializer) const override;
+    void load(Serializer& rSerializer) override;
 
-    void save(Serializer& rSerializer) const override
-    {
-        KRATOS_SERIALIZE_SAVE_BASE_CLASS(rSerializer, LineLoad2DDiffOrderCondition)
-    }
-
-    void load(Serializer& rSerializer) override
-    {
-        KRATOS_SERIALIZE_LOAD_BASE_CLASS(rSerializer, LineLoad2DDiffOrderCondition)
-    }
-
-}; // class LineNormalFluidFlux2DDiffOrderCondition.
+}; // class SurfaceNormalFluidFlux3DDiffOrderCondition.
 
 } // namespace Kratos.
