+++ applications/GeoMechanicsApplication/geo_mechanics_application.h
@@ -28,27 +28,27 @@
 #include "geo_mechanics_application_variables.h"
 
 // conditions
-#include "custom_conditions/Pw_normal_flux_condition.hpp"
-#include "custom_conditions/Pw_point_flux_condition.hpp"
+#include "custom_conditions/Pw_normal_flux_condition.h"
+#include "custom_conditions/Pw_point_flux_condition.h"
 #include "custom_conditions/T_microclimate_flux_condition.h"
 #include "custom_conditions/T_normal_flux_condition.h"
-#include "custom_conditions/U_Pw_face_load_condition.hpp"
-#include "custom_conditions/U_Pw_face_load_interface_condition.hpp"
+#include "custom_conditions/U_Pw_face_load_condition.h"
+#include "custom_conditions/U_Pw_face_load_interface_condition.h"
 #include "custom_conditions/U_Pw_force_condition.hpp"
-#include "custom_conditions/U_Pw_normal_face_load_condition.hpp"
-#include "custom_conditions/U_Pw_normal_flux_FIC_condition.hpp"
-#include "custom_conditions/U_Pw_normal_flux_condition.hpp"
-#include "custom_conditions/U_Pw_normal_flux_interface_condition.hpp"
-#include "custom_conditions/U_Pw_normal_lysmer_absorbing_condition.hpp"
-#include "custom_conditions/axisymmetric_U_Pw_normal_face_load_condition.hpp"
-#include "custom_conditions/axisymmetric_line_normal_fluid_flux_2D_diff_order_condition.hpp"
-#include "custom_conditions/axisymmetric_line_normal_load_2D_diff_order_condition.hpp"
-#include "custom_conditions/line_load_2D_diff_order_condition.hpp"
-#include "custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.hpp"
-#include "custom_conditions/line_normal_load_2D_diff_order_condition.hpp"
-#include "custom_conditions/surface_load_3D_diff_order_condition.hpp"
-#include "custom_conditions/surface_normal_fluid_flux_3D_diff_order_condition.hpp"
-#include "custom_conditions/surface_normal_load_3D_diff_order_condition.hpp"
+#include "custom_conditions/U_Pw_normal_face_load_condition.h"
+#include "custom_conditions/U_Pw_normal_flux_FIC_condition.h"
+#include "custom_conditions/U_Pw_normal_flux_condition.h"
+#include "custom_conditions/U_Pw_normal_flux_interface_condition.h"
+#include "custom_conditions/U_Pw_normal_lysmer_absorbing_condition.h"
+#include "custom_conditions/axisymmetric_U_Pw_normal_face_load_condition.h"
+#include "custom_conditions/axisymmetric_line_normal_fluid_flux_2D_diff_order_condition.h"
+#include "custom_conditions/axisymmetric_line_normal_load_2D_diff_order_condition.h"
+#include "custom_conditions/line_load_2D_diff_order_condition.h"
+#include "custom_conditions/line_normal_fluid_flux_2D_diff_order_condition.h"
+#include "custom_conditions/line_normal_load_2D_diff_order_condition.h"
+#include "custom_conditions/surface_load_3D_diff_order_condition.h"
+#include "custom_conditions/surface_normal_fluid_flux_3D_diff_order_condition.h"
+#include "custom_conditions/surface_normal_load_3D_diff_order_condition.h"
 #include "custom_conditions/thermal_point_flux_condition.h"
 
 // Geometries
