+++ applications/PoromechanicsApplication/poromechanics_application.h
@@ -34,11 +34,11 @@
 #include "custom_conditions/one-phase_flow/U_Pl_normal_liquid_flux_FIC_condition.hpp"
 #include "custom_conditions/one-phase_flow/U_Pl_face_load_interface_condition.hpp"
 #include "custom_conditions/one-phase_flow/U_Pl_normal_liquid_flux_interface_condition.hpp"
-#include "custom_conditions/one-phase_flow/line_load_2D_diff_order_condition.hpp"
-#include "custom_conditions/one-phase_flow/line_normal_load_2D_diff_order_condition.hpp"
+#include "custom_conditions/one-phase_flow/line_load_2D_diff_order_condition.h"
+#include "custom_conditions/one-phase_flow/line_normal_load_2D_diff_order_condition.h"
 #include "custom_conditions/one-phase_flow/line_normal_liquid_flux_2D_diff_order_condition.hpp"
-#include "custom_conditions/one-phase_flow/surface_load_3D_diff_order_condition.hpp"
-#include "custom_conditions/one-phase_flow/surface_normal_load_3D_diff_order_condition.hpp"
+#include "custom_conditions/one-phase_flow/surface_load_3D_diff_order_condition.h"
+#include "custom_conditions/one-phase_flow/surface_normal_load_3D_diff_order_condition.h"
 #include "custom_conditions/one-phase_flow/surface_normal_liquid_flux_3D_diff_order_condition.hpp"
 
 #include "custom_elements/one-phase_flow/U_Pl_small_strain_element.hpp"
